requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0

# AI/ML Libraries
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
textract>=1.6.5
nltk>=3.8.0
spacy>=3.7.0

# Text processing and NLP
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Question generation (specific libraries)
question-generator>=1.0.0

# API clients
requests>=2.31.0
httpx>=0.25.0

# Data handling
pydantic>=2.5.0
jsonlines>=4.0.0
openpyxl>=3.1.0

# Web scraping (for perplexity-like functionality)
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Environment and config
pyyaml>=6.0.0
toml>=0.10.0